requirements.txt
fastapi
uvicorn
qdrant-client
google-generativeai
python-dotenv
markdown
